+++ CHANGELOG.md
@@ -3,14 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.4.10] - 2020-01-02
 ### Added
+- Stats can now be specified per gear type. For instance, this allows the armor stat to be manually set for each piece, rather than using a predefined ratio. Old files will still work as before.
 - Turtle (scute) material. Can be used to craft helmets only.
 - Turtle trait. Mimics the vanilla turtle shell helmet's effect.
 - Netherwood charcoal, burns 2400 ticks (smelts 12 items, configurable)
 - Netherwood charcoal blocks, 10 times the configured burn time
 ### Changed
-- Stats can now be specified per gear type. For instance, this allows the armor stat to be manually set for each piece, rather than using a predefined ratio. Old files will still work as before.
 - Netherite-coated armor now uses the vanilla model texture
 - Updated pt_br.json (SAMUELPV) [#282]
 ### Fixed